Add validation of the address to ParticipantId
The public constructor of ParticipantId has been deprecated and two factory
methods have been added instead. A rudimentary form of address validation is now in use and an InvalidParticipant exception will be thrown when the validation fails using the checked factory method.

The ConsoleClient now uses the new factory method and ensures that a domain name is added for the /add and /remove methods. Note that these use the unsafe factory method.

The server however does not reject deltas with an invalid participant yet as is requested in the original issue. See also Issue 92.